requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==3.0.1
attrs==23.2.0
Automat==22.10.0
beautifulsoup4==4.10.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.12
constantly==23.10.4
cryptography==42.0.7
cssselect==1.2.0
discord.py==1.7.3
filelock==3.14.0
frozenlist==1.4.1
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
itemadapter==0.9.0
itemloaders==1.2.0
jmespath==1.0.1
lxml==5.2.2
multidict==4.7.6
packaging==24.0
parsel==1.9.1
priority==1.3.0
Protego==0.3.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyDispatcher==2.0.7
pyOpenSSL==24.1.0
queuelib==1.7.0
requests==2.26.0
requests-file==2.0.0
Scrapy==2.5.1
service-identity==24.1.0
setuptools==69.5.1
six==1.16.0
soupsieve==2.5
tldextract==5.1.2
Twisted==24.3.0
typing_extensions==4.11.0
urllib3==1.26.18
w3lib==2.1.2
yarl==1.9.4
zope.interface==6.3
